History records can now have NULL usernames, indicating a system operation. LDAPConnection refactored to only be the connection to LDAP server, user operations moved to LDAPUtility.
On LDAP user sign-in current details will be checked against LDAP, and re-synced if there are discrepancies.
LDAP users must be defined by a filter located in Config.  Any username matching the filter will be allowed to log in.
If an account does not exist for them, an account will be created with no roles.

Options have been added into Config for using LDAP groups as MERCURY Roles, this will be added in the future.
LDAP version was moved to Config.